feat: 대시보드에서 제목 스타일 및 관련 코드 정리, 불필요한 코드 제거
